+++ src/utils/array/index.ts
@@ -1,1 +1,4 @@
-export * from "./isUniqueArrayItem/isUniqueArrayItem";
+import isUniqueArrayItem from "./isUniqueArrayItem/isUniqueArrayItem";
+import isUniqueArrayItemWithNewEntry from "./isUniqueArrayItemWithNewEntry/isUniqueArrayItemWithNewEntry";
+
+export { isUniqueArrayItem, isUniqueArrayItemWithNewEntry };