+++ src/utils/array/isUniqueArrayItem/isUniqueArrayItem.ts
@@ -1,14 +1,10 @@
-export const isUniqueArrayItem = (array: string[]): boolean => {
+/**
+ * Check if array items are unique
+ * @param array
+ */
+const isUniqueArrayItem = (array: string[]): boolean => {
   const uniqueArrayItems = new Set([...array]);
   return uniqueArrayItems.size === array.length;
 };
 
-export const isUniqueArrayItemWithNewEntry = (array: string[], entry: string, excludeEntry?: string | false | undefined): boolean => {
-  const arrayNames = [...array, entry];
-
-  if (excludeEntry) {
-    return isUniqueArrayItem(arrayNames.filter((item) => item !== excludeEntry));
-  }
-
-  return isUniqueArrayItem(arrayNames);
-};
+export default isUniqueArrayItem;