+++ src/utils/tree/findNodeByNameInTree/test/findNodeByNameInTree.test.ts
@@ -1,45 +1,45 @@
 import { expect } from "vitest";
-import getNodeByNameInCurrentTree from "@/utils/tree/getNodeByNameInCurrentTree/getNodeByNameInCurrentTree";
+import findNodeByNameInTree from "@/utils/tree/findNodeByNameInTree/findNodeByNameInTree";
 import {
   simpleTree,
   treeWithDecision,
   treeWithTreeMatchTree,
   treeWithTreeMock,
   treeWithTreeNoMatch,
-} from "@/utils/tree/getNodeByNameInCurrentTree/test/mock";
+} from "@/utils/tree/findNodeByNameInTree/test/mock";
 
 describe("getNodeNames", () => {
   test("Simple tree", () => {
     const { input, output, searchName } = simpleTree;
-    const result = getNodeByNameInCurrentTree(input, searchName);
+    const result = findNodeByNameInTree(input, searchName);
 
     expect(result).toEqual(output);
   });
 
   test("tree with decision", () => {
     const { input, output, searchName } = treeWithDecision;
-    const result = getNodeByNameInCurrentTree(input, searchName);
+    const result = findNodeByNameInTree(input, searchName);
 
     expect(result).toEqual(output);
   });
 
   test("tree with tree (he searching just in current tree)", () => {
     const { input, output, searchName } = treeWithTreeMock;
-    const result = getNodeByNameInCurrentTree(input, searchName);
+    const result = findNodeByNameInTree(input, searchName);
 
     expect(result).toEqual(output);
   });
 
   test("tree with tree (search in last tree return null)", () => {
     const { input, output, searchName } = treeWithTreeNoMatch;
-    const result = getNodeByNameInCurrentTree(input, searchName);
+    const result = findNodeByNameInTree(input, searchName);
 
     expect(result).toEqual(output);
   });
 
   test("tree with tree (matching tree name)", () => {
     const { input, output, searchName } = treeWithTreeMatchTree;
-    const result = getNodeByNameInCurrentTree(input, searchName);
+    const result = findNodeByNameInTree(input, searchName);
 
     expect(result).toEqual(output);
   });