+++ src/utils/tree/removeNode/removeNode.ts
@@ -1,6 +1,6 @@
 import type { TreeNode } from "@/features/Treege/type/TreeNode";
+import findParentNodeByNameInTree from "@/utils/tree/findParentNodeByNameInTree/findParentNodeByNameInTree";
 import getNode from "@/utils/tree/getNode/getNode";
-import getParentNodeByNameInCurrentTree from "@/utils/tree/getParentNodeByNameInCurrentTree/getParentNodeByNameInCurrentTree";
 import getTree from "@/utils/tree/getTree/getTree";
 
 interface RemoveNodeParams {
@@ -9,8 +9,15 @@ interface RemoveNodeParams {
   name: string;
 }
 
+/**
+ * Remove tree node
+ * @param parent
+ * @param node
+ */
 const removeTreeNode = (parent: TreeNode | null, node: TreeNode | null) => {
-  if (!parent || !node) return null;
+  if (!parent || !node) {
+    return null;
+  }
 
   Object.defineProperty(parent, "children", { value: parent.children.filter((child) => child.name !== node.name) });
   Object.defineProperty(parent, "attributes", { value: { ...parent.attributes, isLeaf: true } });
@@ -18,11 +25,23 @@ const removeTreeNode = (parent: TreeNode | null, node: TreeNode | null) => {
   return null;
 };
 
+/**
+ * Get parent node by name in current tree
+ * @param tree
+ * @param path
+ * @param name
+ */
 const getParentTreeNode = (tree: TreeNode, path: string, name: string) => {
   const currentTree = getTree(tree, path);
-  return getParentNodeByNameInCurrentTree(currentTree, name);
+  return findParentNodeByNameInTree(currentTree, name);
 };
 
+/**
+ * Remove node from tree
+ * @param tree
+ * @param path
+ * @param name
+ */
 const removeNode = ({ tree, path, name }: RemoveNodeParams) => {
   const treeCopy = structuredClone(tree);