+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # [Versions](https://github.com/Tracktor/treege/releases)
 
-## v0.11.1
-- **[build]** : Improve build process
+## v0.11.2
+- **[enhancement]** : Rename`getNodeByNameInCurrentTree`function to `findNodeByNameInTree` and improve it
+- **[enhancement]** : Rename `getParentNodeByNameInCurrentTree` function to `findParentNodeByNameInTree`   and improve it
+- **[enhancement]** : Improve`getNodeNames` function
+- **[enhancement]** : Improve`getTree` function
+- **[enhancement]** : Rename `updateNodeInTree` function to`updateNodeInTree` function and improve it
+- **[refactor]** : Split `isUniqueArrayItem` and `isUniqueArrayItemWithNewEntry`